% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_rows_have_missing}
\alias{df_rows_have_missing}
\title{Check if rows of a data frame have missing values}
\usage{
df_rows_have_missing(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A logical vector of length \code{nrow(df)} indicating which rows have
missing values.
}
\description{
Check if rows of a data frame have missing values
}
\examples{
df <- data.frame(x = c(1, 2, NA, 4), y = c(NA, 2, 3, 4))
df_rows_have_missing(df)
}
