% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_likeliness.R
\name{distribution_likeliness}
\alias{distribution_likeliness}
\title{Likeliness of a distribution to produce a specified value}
\usage{
distribution_likeliness(distributions, value)
}
\arguments{
\item{distributions}{A list of distributions.}

\item{value}{The value to calculate the likeliness for.}
}
\value{
A vector of likeliness values, one for each distribution.
}
\description{
For a set of distributions, calculate the likeliness of each
distribution to produce a specified value.
}
\details{
The likeliness is calculated as the density of the distribution at the
specified value divided by the sum of the densities of all distributions.

If a density evaluates to \code{NA}, it is ignored in the sum.
}
\examples{
distributions <- list(
  distionary::dst_normal(mean = 0, sd = 1),
  distionary::dst_normal(mean = 1, sd = 1),
  distionary::dst_null()
)
distribution_likeliness(distributions, 0)
}
