% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_diagnostics.R
\name{dl_diagnostics_basename}
\alias{dl_diagnostics_basename}
\title{Default diagnostics RDS basename (under \verb{derived/})}
\usage{
dl_diagnostics_basename()
}
\description{
Default diagnostics RDS basename (under \verb{derived/})
}
