% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_null.R
\name{dl_null}
\alias{dl_null}
\title{Make a null distributional learning object}
\usage{
dl_null(data = NULL)
}
\arguments{
\item{data}{A data frame of training data. Useful if you want
to run \code{predict()} without specifying \code{newdata},
in which case the number of rows of the data frame will be used.}
}
\value{
A distributional learning object with subclass
"dl_null".
}
\description{
Useful for when a distributional learning method fails
}
\examples{
dl_null()
}
