% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_rqforest.R
\name{dl_rqforest}
\alias{dl_rqforest}
\title{Make a quantile regression forest distributional learning model}
\usage{
dl_rqforest(
  data,
  yname,
  xnames,
  ...,
  na_action = c("drop", "null"),
  min_obs = 5
)
}
\arguments{
\item{data}{A data frame.}

\item{yname}{The name of the response variable.}

\item{xnames}{The names of the predictor variables.}

\item{...}{Other arguments passed down to \code{quantregForest::quantregForest()}.}

\item{na_action}{The action to take if there are missing values in the data;
either "drop" to drop rows with missing values, or "null" to return a
"dl_null" object.}

\item{min_obs}{The minimum number of observations required to fit the model.
If the number of observations is less than \code{min_obs}, a "dl_null" object
is returned.}
}
\value{
A distributional learning model with subclass "dl_rqforest".
}
\description{
Make a quantile regression forest distributional learning model
}
\examples{
set.seed(123)
df <- tibble::tibble(x = rnorm(100), y = x + rnorm(100))
dl_rqforest(data = df, yname = "y", xnames = "x")
}
