% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ecdf_to_dst}
\alias{ecdf_to_dst}
\title{Convert an ecdf to an empirical distribution}
\usage{
ecdf_to_dst(fun)
}
\arguments{
\item{fun}{An ecdf function.}
}
\value{
An empirical distribution from the distionary package.
}
\description{
Convert an ecdf to an empirical distribution
}
\examples{
fun <- ecdf(rnorm(100))
ecdf_to_dst(fun)
}
