% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivers.R
\name{eval_drivers_density}
\alias{eval_drivers_density}
\title{Evaluate the drivers density}
\usage{
eval_drivers_density(object, newdata, given = NULL)
}
\arguments{
\item{object}{A \code{"drivers"} object.}

\item{newdata}{Data frame with the predictor columns.}

\item{given}{For two predictors, optionally the name of the predictor to
condition on: returns the conditional density of the other predictor.}
}
\value{
Numeric vector of density values.
}
\description{
Evaluate the drivers density
}
