% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_empirical_to_gp.R
\name{fit_and_graft_gp}
\alias{fit_and_graft_gp}
\title{Fit and Graft a GPD to the upper tail of an empirical distribution.}
\usage{
fit_and_graft_gp(dst, adaptive_threshold)
}
\arguments{
\item{dst}{An empirical distribution: the object returned by
\code{distionary::dst_empirical()}.}

\item{adaptive_threshold}{Number between 0 and 1. Takes the smallest
value between the quantile of \code{dst} (with the built-in weights)
and the quantile as if all weights were equal. (This prevents
situations where the empirical tail is dominated by a few outcomes.)}
}
\value{
A distionary distribution with the GPD grafted back on.
}
\description{
Fit and Graft a GPD to the upper tail of an empirical distribution.
}
\note{
Requires distplyr > 0.2.0; may have to use the development version.
}
