% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pot_events.R
\name{get_pot_events}
\alias{get_pot_events}
\title{Get Peaks Over Threshold Events}
\usage{
get_pot_events(
  data,
  threshold,
  ...,
  flow_col = "flow",
  date_col = "date",
  min_gap = 0
)
}
\arguments{
\item{data}{A data frame containing the flow and date columns.}

\item{threshold}{The threshold value; either a function that takes a vector
of flow values and returns a single threshold value, or a single numeric
threshold value.}

\item{...}{Additional arguments, currently unused.}

\item{flow_col}{The name of the flow column (character).}

\item{date_col}{The name of the date column (character).}

\item{min_gap}{The smallest number of observations in between two consecutive
"runs" (i.e., contiguous threshold exceedances) to be considered a single
event. Non-negative numeric. If 1, no merging is done, because each run
will always have at least one non-exceedance between them.}
}
\value{
The original data frame (as a tibble), filtered to the highest
flow value for each event. If the max flow is sustained for more than one
row, the first row is returned.

The threshold value is also available, stored as an attribute of the returned
data frame, named "threshold".
}
\description{
Identify events where the flow exceeds a given threshold.
}
\examples{
data <- tibble::tibble(
  date = as.Date(paste0("2020-01-", 1:30)),
  flow = c(
    1, 2, 3, 4, 5,
    5, 3, 2, 1, 0,
    1, 2, 5, 4, 4,
    5, 3, 2, 1, 0,
    1, 2, 3, 4, 5,
    4, 3, 2, 1, 0
  )
)
plot(data$flow, type = "l")
get_pot_events(data, threshold = 2)

## merge runs separated by fewer than three observations:
x <- tibble::tibble(
  date = lubridate::ymd_hms("2020-01-01 00:00:00") +
    lubridate::dhours(6) * 0:12,
  flow = c(4, 4, 0, 5, 5, 0, 0, 2, 0, 0, 0, 6, 6)
)
ggplot2::ggplot(x, ggplot2::aes(date, flow)) +
  ggplot2::geom_line() +
  ggplot2::geom_point()
get_pot_events(x, threshold = 1, min_gap = 1)
get_pot_events(x, threshold = 1, min_gap = 2)
get_pot_events(x, threshold = 1, min_gap = 3)
get_pot_events(x, threshold = 1, min_gap = 4)
}
