% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_likeliness.R
\name{grid_from_scatter}
\alias{grid_from_scatter}
\title{Regular grid from a scatter}
\usage{
grid_from_scatter(x, y, data = NULL, size = 30, mult = 1)
}
\arguments{
\item{x, y}{Names of the x and y variables (unquoted); they will be evaluated
in the \code{data} data mask.}

\item{data}{A possible data frame to evaluate the x and y variables in; NULL
if not needed (i.e., grab vectors from the enclosing environment).}

\item{size}{The number of grid points in each dimension; vector of length 1
for the same size in both dimensions, or vector of length 2 corresponding
to (x, y).}

\item{mult}{A multiplier for the range of each variable; vector of length 1
for the same size in both dimensions, or vector of length 2 corresponding
to (x, y). NOTE: expansion only happens in the positive direction (bottom
left corner stays the same) because the focus is on extremes.}
}
\value{
A data frame with columns \code{x} and \code{y}.
}
\description{
For a scattering of x-y points, this function makes a regular x-y grid
based on the range of each variable.
}
\examples{
set.seed(123)
x <- rnorm(100)
y <- rnorm(100)
grid <- grid_from_scatter(x, y, size = 10, mult = 1.1)
plot(grid$x, grid$y)
points(x, y, col = "red", pch = 19)
legend(
  "bottomright",
  legend = c("Grid", "Scatter"),
  col = c("black", "red"),
  pch = c(1, 19)
)
}
