% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{list_analyses}
\alias{list_analyses}
\title{List the analyses in the repository}
\usage{
list_analyses(root = here::here())
}
\arguments{
\item{root}{Repository root.}
}
\value{
A character vector of analysis names (folders containing an
\code{analysis.yaml}).
}
\description{
List the analyses in the repository
}
