% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_tail.R
\name{mixture_tail_survival}
\alias{mixture_tail_survival}
\title{Survival function of a mixture tail}
\usage{
mixture_tail_survival(mt, x)
}
\arguments{
\item{mt}{A \code{mixture_tail}.}

\item{x}{Numeric vector of values.}
}
\value{
Numeric vector of exceedance probabilities. Values below
\code{\link[=mixture_tail_lower]{mixture_tail_lower()}} return \code{NA} -- the mixture has empirical mass there
that this object does not carry.
}
\description{
Survival function of a mixture tail
}
\examples{
mt <- mixture_tail(c(10, 12), c(0.5, 0.5), c(2, 3), 0.15)
mixture_tail_survival(mt, c(20, 100))
}
