% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_dstlrn.R
\name{new_dstlrn}
\alias{new_dstlrn}
\title{Make a distributional learning object}
\usage{
new_dstlrn(x, ..., subclass = character())
}
\arguments{
\item{...}{Other arguments to add as attributes to the object.}

\item{subclass}{The subclass of the distributional learning object.}

\item{l}{A list to use as the basis for the object.}
}
\value{
A distributional learning object with subclass \code{subclass}.
}
\description{
Make a distributional learning object
}
\examples{
new_dstlrn(list())
new_dstlrn(list(a = 1, b = 2), subclass = "my_subclass")
}
