% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.dstlrn}
\alias{print.dstlrn}
\title{Print a distributional learning object}
\usage{
\method{print}{dstlrn}(x, ...)
}
\arguments{
\item{x}{A distributional learning object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a distributional learning object
}
