% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_rain_snow.R
\name{print.joint_rain_snow}
\alias{print.joint_rain_snow}
\title{Print a \code{joint_rain_snow} object}
\usage{
\method{print}{joint_rain_snow}(x, ...)
}
\arguments{
\item{x}{A \code{joint_rain_snow} object from \code{\link[=fit_joint_rain_snow_cell]{fit_joint_rain_snow_cell()}}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{joint_rain_snow} object
}
