% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rp_reporting}
\alias{rp_reporting}
\title{Return periods for reporting}
\usage{
rp_reporting()
}
\value{
A vector of commonly reported return periods.
}
\description{
Return periods for reporting
}
\examples{
rp_reporting()
}
