% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport_lab.R
\name{td_x_quantile}
\alias{td_x_quantile}
\alias{td_x_cdf}
\title{The covariate distribution}
\usage{
td_x_quantile(dgp, p)

td_x_cdf(dgp, x)
}
\arguments{
\item{dgp}{A \code{transport_dgp}.}

\item{p}{Numeric vector of probabilities.}

\item{x}{Numeric vector of covariate values.}
}
\value{
A numeric vector.
}
\description{
The covariate distribution
}
\examples{
td_x_quantile(transport_dgp(), c(0.1, 0.9))
}
