% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_process_dgp.R
\name{tp_marginal_quantile}
\alias{tp_marginal_quantile}
\title{Marginal quantile of the annual peak}
\usage{
tp_marginal_quantile(dgp, s)
}
\arguments{
\item{dgp}{A \code{two_process_dgp}.}

\item{s}{Numeric vector of upper-tail probabilities. For annual maxima the
\code{T}-year level is \code{s = 1 / T}.}
}
\value{
A numeric vector.
}
\description{
Marginal quantile of the annual peak
}
\examples{
tp_marginal_quantile(two_process_dgp(), 1 / c(20, 100, 200))
}
