% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_process_dgp.R
\name{tp_x_quantile}
\alias{tp_x_quantile}
\alias{tp_x_cdf}
\title{The rainfall distribution}
\usage{
tp_x_quantile(dgp, p)

tp_x_cdf(dgp, x)
}
\arguments{
\item{dgp}{A \code{two_process_dgp}.}

\item{p}{Numeric vector of probabilities.}

\item{x}{Numeric vector of rainfall values.}
}
\value{
A numeric vector.
}
\description{
The rainfall distribution
}
\examples{
tp_x_quantile(two_process_dgp(), c(0.1, 0.9))
}
