% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_null.R, R/dl_rqforest.R, R/predict.R
\name{predict.dl_null}
\alias{predict.dl_null}
\alias{predict.dl_rqforest}
\alias{predict.dstlrn}
\title{Predict from a distributional learning model}
\usage{
\method{predict}{dl_null}(object, newdata, ...)

\method{predict}{dl_rqforest}(object, newdata = NULL, ...)

\method{predict}{dstlrn}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A distributional learning object.}

\item{newdata}{A data frame of new data to predict. If \code{NULL}, the training
data is used.}

\item{...}{Additional arguments passed on to specific methods.}
}
\value{
A list of probability distributions, one for each row of \code{newdata}.
}
\description{
\code{predict()} is an S3 generic from the stats package. Methods are provided
for distributional learning objects inheriting from class \code{"dstlrn"}.
}
\section{Functions}{
\itemize{
\item \code{predict(dl_null)}: Predict from a null distributional learning model.

\item \code{predict(dl_rqforest)}: Predict from a quantile regression forest
distributional learning model.

}}
